Rename MACH Tutorial wing to Simple Transonic Wing (STW)
