+++ AircraftSpecs/README.md
@@ -2,5 +2,5 @@
 
 ## File descriptions:
 
-- **`MACHTutorialWingSpecs.py`:** Python file containing various aircraft and mission specifications, designed to be imported and used in other scripts, particularly the `aircraftSpecs` dictionary.
-- **`MACHTutorialWingFlightPoints.py`:** Python file defining the flight point objects for each of the flight points used in the MACH Tutorial Wing benchmark problems.
+- **`STWSpecs.py`:** Python file containing various aircraft and mission specifications, designed to be imported and used in other scripts, particularly the `aircraftSpecs` dictionary.
+- **`STWFlightPoints.py`:** Python file defining the flight point objects for each of the flight points used in the MACH Tutorial Wing benchmark problems.
